+++ compiler/src/prog8/compiler/astprocessing/AstPreprocessor.kt
@@ -447,15 +447,15 @@ class AstPreprocessor(val program: Program,
                         val replacement = if(tgt2!=null) {
                             if(tgt2 is BuiltinFunctionPlaceholder) {
                                 val unscopedTarget = IdentifierReference(listOf(tgt2.name), alias.position)
-                                Alias(alias.alias, unscopedTarget, alias.isPrivate, alias.position)
+                                Alias(alias.alias, unscopedTarget, alias.visibility, alias.position)
                             } else if(tgt2.scopedName != chainedTargetName.nameInSource) {
                                 val scopedTarget = IdentifierReference(tgt2.scopedName, alias.position)
-                                Alias(alias.alias, scopedTarget, alias.isPrivate, alias.position)
+                                Alias(alias.alias, scopedTarget, alias.visibility, alias.position)
                             } else {
-                                Alias(alias.alias, chainedTargetName, alias.isPrivate, alias.position)
+                                Alias(alias.alias, chainedTargetName, alias.visibility, alias.position)
                             }
                         } else {
-                            Alias(alias.alias, chainedTargetName, alias.isPrivate, alias.position)
+                            Alias(alias.alias, chainedTargetName, alias.visibility, alias.position)
                         }
                         return listOf(AstReplaceNode(alias, replacement, parent))
                     }
@@ -561,7 +561,7 @@ class AstPreprocessor(val program: Program,
                 .names(membername)
                 .type(VarDeclType.CONST)
                 .value(membervalue)
-                .isPrivate(enum.isPrivate)
+                .visibility(enum.visibility)
                 .build()
         }